Send request_dropped error reply directly from do_work.
When a call is dropped because there are no workers available and the
queue is full, send the request_dropped reply directly from do_work.
This fixes a bug where the request that is dropped is the oldest request
in the queue, but previously the request_dropped reply is sent to the
request that was just received.

This is also more consistent with how the request_timeout reply is sent
and will allow some refactoring in the future.