When handling a request timeout don't lose worker and try next request.
When a request is timed out because it is too old, the worker that was
taken from the available worker stack to handle the request was
forgotten.

In addition, do_work just returned even though there might be other
requests on the queue; now it tries the next request.

Finally, there was not accounting for these timeout-out requests.  They
are now counted as dropped tasks.