Enhance worker lifetime handling.
Previously, max_worker_age_ms could only be specified as a lifetime in
milliseconds or 'infinity'.  Now it can also be specified as
{LifetimeMS, JitterMS}, and {Module, Function, Args}.  Adding jitter to
the worker lifetimes can be useful when the worker's initialization is
expensive and you want to avoid having lots of workers restart at the
same time.

To support each workers having a different lifetime, its end time is now
calculated when the worker is started and, instead of the start time,
this end time is carried around with the worker's state.